Create README.md
The objective of this report is to present a detailed analysis of policy gradient methods
and their variants based on a review of recent scientific articles. The work conducted
during this research internship involved reading, analyzing, and synthesizing key academic
publications in this domain to gain a deeper understanding of the theoretical foundations
and current advancements. While this report does not contain original experimental
results, it aims to provide an overview of policy gradient techniques and discuss their
advantages, limitations, and future research perspectives.